feat(npwd/db): move database files to @npwd/database
