+++ apps/game/server/match/match.service.ts
@@ -1,5 +1,5 @@
 import { formatMatches, formatProfile, matchLogger } from './match.utils';
-import MatchDB, { _MatchDB } from './match.db';
+import { MatchDB, _MatchDB } from '@npwd/database';
 import {
   FormattedMatch,
   FormattedProfile,
