+++ packages/database/src/index.ts
@@ -1,2 +1,12 @@
 export * from './contacts/index';
+export * from './notes/index';
+export * from './twitter/index';
+export * from './players/index';
+export * from './photo/index';
+export * from './messages/index';
+export * from './match/index';
+export * from './marketplace/index';
+export * from './darkchat/index';
+export * from './calls/index';
+
 export { pool } from './db/pool';
