+++ packages/database/src/match/index.ts
@@ -1,9 +1,9 @@
-import { Like, Match, NewProfile, Profile } from '../../../typings/match';
+import { Like, Match, NewProfile, Profile } from '@typings/match';
 import { pool } from '../db/pool';
 import { ResultSetHeader } from 'mysql2';
-import { config } from '../config';
-import { generateProfileName } from '../utils/generateProfileName';
-import { matchLogger } from './match.utils';
+import { config } from '@npwd/config/server';
+import { generateProfileName } from '@game/server/utils/generateProfileName';
+import { matchLogger } from '@game/server/match/match.utils';
 import DbInterface from '../db/db_wrapper';
 
 const DEFAULT_IMAGE = 'https://upload.wikimedia.org/wikipedia/commons/a/ac/No_image_available.svg';
@@ -240,6 +240,4 @@ export class _MatchDB {
   }
 }
 
-const MatchDB = new _MatchDB();
-
-export default MatchDB;
+export const MatchDB = new _MatchDB();
